% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CmdstanrBrmsImage-package.R
\docType{package}
\name{CmdstanrBrmsImage-package}
\alias{CmdstanrBrmsImage}
\alias{CmdstanrBrmsImage-package}
\title{CmdstanrBrmsImage: Make the HPC Like Bayesian Statistics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: GW McElfresh \email{mcelfreshgw@gmail.com}

}
\keyword{internal}
