% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAE.R
\name{hurdleGaussianLoss}
\alias{hurdleGaussianLoss}
\title{Calculate Loss for Hurdle Gaussian model}
\usage{
hurdleGaussianLoss(pZero, positiveParams, rawData)
}
\arguments{
\item{pZero}{Bernoulli probability for zero-inflation component.}

\item{positiveParams}{List containing the mean and log-variance for the Gaussian component.}

\item{rawData}{output data from the model's forward pass}
}
\description{
This module calculates the loss for a VAE with a hurdle Gaussian decoder.
}
