% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConstructors.R
\name{scrnaseqDataLoader}
\alias{scrnaseqDataLoader}
\title{Single Cell RNASeq Dataset Loader}
\usage{
scrnaseqDataLoader(seuratObj, layer = "data")
}
\arguments{
\item{data}{A matrix of single cell RNASeq data.}
}
\value{
A dataset loader for single cell RNASeq data.
}
\description{
This module defines a dataset loader for single cell RNASeq data.
}
