% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAE.R
\name{VAE_HurdleGaussian}
\alias{VAE_HurdleGaussian}
\title{Variational Autoencoder (VAE) with Hurdle Gaussian Decoder}
\usage{
VAE_HurdleGaussian(input_dim, latent_dim)
}
\arguments{
\item{input_dim}{The number of input features.}

\item{latent_dim}{The number of latent dimensions.}
}
\value{
A VAE module with a hurdle Gaussian decoder.
}
\description{
This module defines a VAE with a hurdle Gaussian decoder.
}
