% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConstructors.R
\name{scrnaseqDataLoader}
\alias{scrnaseqDataLoader}
\title{Single Cell RNASeq Dataset Loader}
\usage{
scrnaseqDataLoader(seuratObj, assay = "RNA", layer = "data")
}
\arguments{
\item{seuratObj}{Seurat Object containing single cell RNASeq data.}

\item{assay}{The name of the assay containing the data.}

\item{layer}{Seurat object's layer.}
}
\value{
A dataset loader for single cell RNASeq data.
}
\description{
This module defines a dataset loader for single cell RNASeq data.
}
