% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WriteSeuratToAnnData}
\alias{WriteSeuratToAnnData}
\title{WriteToAnnData}
\usage{
WriteSeuratToAnnData(
  seuratObj,
  outputPath = "./seuratToAnnDataOutput",
  assayName = "RNA",
  layer = "counts",
  overwrite = FALSE
)
}
\arguments{
\item{seuratObj}{Seurat object to write.}

\item{overwrite}{Overwrite the file if it already exists.}

\item{path}{Path to write the Seurat object to.}
}
\description{
Write Seurat object to an AnnData-digestable format.
}
